// src/interfaces/model/apiObject.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ApiAuthenticationObject } from './apiAuthenticationObject'
import { LimitObject } from './limitObject'
import { VersionsDataObject } from './versionsDataObject'

/**
 * The Api Management DTO corresponding ApiDefinition Tyk object
 */
export interface ApiObject {
  apiAuthentication: ApiAuthenticationObject
  /**
   * The unique API identifier.  It is recommended to make this a UUID.
   */
  apiId: string
  internal?: boolean
  limits?: LimitObject
  /**
   * The path to listen on, e.g. /api or /. Any requests coming into the host, on the port that Tyk is configured to run on, that go to this path will have the rules defined in the API Definition applied.
   */
  listenPath: string
  /**
   * Human readable name of the API.
   */
  name: string
  /**
   * This defines the target URL that the request should be proxied to if it passes all checks in Tyk.
   */
  targetUrl: string
  versionData: VersionsDataObject
}